Changed doomed's Dark Vision to only add damage bonus inside Creeping Darkness Changed hate lost per turn (slower at higher levels of hate, faster at lower levels) Fixed some line of sight issues with doomed's Feed Added cursed Strife talent tree Added doomed Primal Magic talent tree Added cursed/doomed Fateful Aura talent tree Improved a few cursed/doomed particle effects for smaller tiles Balance adjustments in cursed/doomed trees Fixed NPC "The Mouth" to support pinning correctly
git-svn-id: http://svn.net-core.org/repos/t-engine4@3831 51575b47-30f0-44d4-a5cc-537603b46e54

+++ game/modules/tome/data/gfx/particles/force_hit.lua
@@ -28,7 +28,7 @@ local first = true
 
 return { generator = function()
 	local life = rng.range(3, 8)
-	local size = rng.float(2, 4)
+	local size = rng.float(3, 5)
 	local alpha = rng.range(80, 160) / 255
 	local moveAngle = angle + math.rad(rng.range(-25, 25))
 	local offsetDistance = engine.Map.tile_w * 0.7
